+++ lib/okf.rb
@@ -40,6 +40,8 @@ module OKF
   require "okf/bundle"
   require "okf/bundle/graph"
   require "okf/bundle/search"
+  require "okf/bundle/search/index"
+  require "okf/bundle/search/scan"
   require "okf/bundle/validator"
   require "okf/bundle/validator/result"
   require "okf/bundle/linter"